+++ components/editor/NewsletterEditor.js
@@ -1,14 +1,13 @@
 import React, { Component } from 'react'
 import PropTypes from 'prop-types'
-import { Document as SlateDocument } from 'slate'
 import { Editor as SlateEditor } from 'slate-react'
 import { css } from 'glamor'
 
-import MarkdownSerializer from '../../lib/serializer'
-import addValidation, { findOrCreate } from './utils/serializationValidation'
 import styles from './styles'
 import Sidebar from './Sidebar'
 import MetaData from './modules/meta/ui'
+
+import createDocumentModule from './modules/document'
 
 import marks, {
   BoldButton,
@@ -64,170 +63,29 @@ const newsletterStyles = {
   maxWidth: 'calc(100vw - 190px)'
 }
 
-const autoMeta = documentNode => {
-  const data = documentNode.data
-  const autoMeta = !data || !data.size || data.get('auto')
-  if (!autoMeta) {
-    return null
-  }
-  const cover = documentNode.nodes
-    .find(n => n.type === COVER && n.kind === 'block')
-  if (!cover) {
-    return null
-  }
-
-  const title = cover.nodes.first()
-  const lead = cover.nodes.get(1)
-
-  const newData = data
-    .set('auto', true)
-    .set('title', title ? title.text : '')
-    .set('description', lead ? lead.text : '')
-    .set('image', cover.data.get('src'))
-
-  return data.equals(newData)
-    ? null
-    : newData
-}
-
-const documentRule = {
-  match: object => object.kind === 'document',
-  matchMdast: node => node.type === 'root',
-  fromMdast: (node, index, parent, visitChildren) => {
-    const cover = findOrCreate(node.children, {
-      type: 'zone', identifier: COVER
-    }, {
-      children: []
-    })
-
-    let center = findOrCreate(node.children, {
-      type: 'zone', identifier: CENTER
-    }, {
-      children: []
-    })
-
-    const centerIndex = node.children.indexOf(center)
-    const before = []
-    const after = []
-    node.children.forEach((child, index) => {
-      if (child !== cover && child !== center) {
-        if (index > centerIndex) {
-          after.push(child)
-        } else {
-          before.push(child)
-        }
+const documentModule = createDocumentModule({
+  TYPE: 'document',
+  subModules: [
+    {
+      TYPE: COVER,
+      helpers: {
+        serializer: coverSerializer
       }
-    })
-    if (before.length || after.length) {
-      center = {
-        ...center,
-        children: [
-          ...before,
-          ...center.children,
-          ...after
-        ]
+    },
+    {
+      TYPE: CENTER,
+      helpers: {
+        serializer: centerSerializer
       }
     }
-
-    const documentNode = {
-      data: node.meta,
-      kind: 'document',
-      nodes: [
-        coverSerializer.fromMdast(cover),
-        centerSerializer.fromMdast(center)
-      ]
-    }
-
-    const newData = autoMeta(
-      SlateDocument.fromJSON(documentNode)
-    )
-    if (newData) {
-      documentNode.data = newData.toJS()
-    }
-
-    return {
-      document: documentNode,
-      kind: 'state'
-    }
-  },
-  toMdast: (object, index, parent, visitChildren, context) => {
-    const firstNode = object.nodes[0]
-    if (!firstNode || firstNode.type !== COVER || firstNode.kind !== 'block') {
-      context.dirty = true
-    }
-    const secondNode = object.nodes[1]
-    if (!secondNode || secondNode.type !== CENTER || secondNode.kind !== 'block') {
-      context.dirty = true
-    }
-    if (object.nodes.length !== 2) {
-      context.dirty = true
-    }
-
-    const cover = findOrCreate(object.nodes, { kind: 'block', type: COVER })
-    const center = findOrCreate(
-      object.nodes,
-      { kind: 'block', type: CENTER },
-      { nodes: [] }
-    )
-    const centerIndex = object.nodes.indexOf(center)
-    object.nodes.forEach((node, index) => {
-      if (node !== cover && node !== center) {
-        center.nodes[index > centerIndex ? 'push' : 'unshift'](node)
-      }
-    })
-    return {
-      type: 'root',
-      meta: object.data,
-      children: [
-        coverSerializer.toMdast(cover),
-        centerSerializer.toMdast(center)
-      ]
-    }
-  }
-}
-
-export const serializer = new MarkdownSerializer({
-  rules: [
-    documentRule
   ]
 })
 
-export const newDocument = ({title}) => serializer.deserialize(
-`<section><h6>${COVER}</h6>
-
-# ${title}
-
-<hr/></section>
-
-<section><h6>${CENTER}</h6>
-
-Ladies and Gentlemen,
-
-<hr/></section>
-`)
-
-addValidation(documentRule, serializer, 'document')
-
-const documentPlugin = {
-  schema: {
-    rules: [
-      documentRule
-    ]
-  },
-  onBeforeChange: (change) => {
-    const newData = autoMeta(change.state.document)
-
-    if (newData) {
-      change.setNodeByKey(change.state.document.key, {
-        data: newData
-      })
-      return change
-    }
-  }
-}
+export const serializer = documentModule.helpers.serializer
+export const newDocument = documentModule.helpers.newDocument
 
 const plugins = [
-  documentPlugin,
+  ...documentModule.plugins,
   ...marks.plugins,
   ...headlines.plugins,
   ...lead.plugins,
